+++ irs-api/src/test/java/org/eclipse/tractusx/irs/aaswrapper/job/delegate/RelationshipDelegateTest.java
@@ -39,6 +39,7 @@ import java.util.List;
 
 import org.eclipse.tractusx.irs.aaswrapper.job.AASTransferProcess;
 import org.eclipse.tractusx.irs.aaswrapper.job.ItemContainer;
+import org.eclipse.tractusx.irs.component.PartChainIdentificationKey;
 import org.eclipse.tractusx.irs.component.enums.ProcessStep;
 import org.eclipse.tractusx.irs.edc.client.EdcSubmodelFacade;
 import org.eclipse.tractusx.irs.edc.client.exceptions.EdcClientException;
@@ -73,12 +74,31 @@ class RelationshipDelegateTest {
 
         // when
         final ItemContainer result = relationshipDelegate.process(itemContainerWithShell, jobParameter(),
-                aasTransferProcess, "itemId");
+                aasTransferProcess, createKey());
 
         // then
         assertThat(result).isNotNull();
         assertThat(result.getRelationships()).isNotEmpty();
         assertThat(aasTransferProcess.getIdsToProcess()).isNotEmpty();
+    }
+
+    @Test
+    void shouldPutTombstoneForMissingBpn() throws EdcClientException {
+        final ItemContainer.ItemContainerBuilder itemContainerWithShell = ItemContainer.builder()
+                                                                                       .shell(shellDescriptor(
+                                                                                               List.of(submodelDescriptor(
+                                                                                                       singleLevelBomAsBuiltAspectName,
+                                                                                                       "address"))));
+        // when
+        final ItemContainer result = relationshipDelegate.process(itemContainerWithShell, jobParameter(),
+                new AASTransferProcess(), PartChainIdentificationKey.builder().globalAssetId("testId").build());
+
+        // then
+        assertThat(result).isNotNull();
+        assertThat(result.getTombstones()).hasSize(1);
+        assertThat(result.getTombstones().get(0).getCatenaXId()).isEqualTo("testId");
+        assertThat(result.getTombstones().get(0).getProcessingError().getProcessStep()).isEqualTo(
+                ProcessStep.SUBMODEL_REQUEST);
     }
 
     @Test
@@ -96,7 +116,7 @@ class RelationshipDelegateTest {
 
         // when
         final ItemContainer result = relationshipDelegate.process(itemContainerWithShell, jobParameter(),
-                new AASTransferProcess(), "itemId");
+                new AASTransferProcess(), createKey());
 
         // then
         assertThat(result).isNotNull();
@@ -120,7 +140,7 @@ class RelationshipDelegateTest {
 
         // when
         final ItemContainer result = relationshipDelegate.process(itemContainerWithShell, jobParameter(),
-                new AASTransferProcess(), "itemId");
+                new AASTransferProcess(), createKey());
 
         // then
         assertThat(result).isNotNull();
@@ -128,6 +148,10 @@ class RelationshipDelegateTest {
         assertThat(result.getTombstones().get(0).getCatenaXId()).isEqualTo("itemId");
         assertThat(result.getTombstones().get(0).getProcessingError().getProcessStep()).isEqualTo(
                 ProcessStep.SUBMODEL_REQUEST);
+    }
+
+    private static PartChainIdentificationKey createKey() {
+        return PartChainIdentificationKey.builder().globalAssetId("itemId").bpn("bpn123").build();
     }
 
 }